+++ EnableLidWake/EnableLidWake.cpp
@@ -131,11 +131,10 @@ void LWEnabler::processKext(KernelPatcher& patcher, size_t index, mach_vm_addres
     for (size_t i = 0; i < kextListSize; i++)
     {
         if (kextList[i].loadIndex != index) continue;
-        //
+
         // Enable lid wake for Haswell (Azul) platform
-        //
-        if (!(progressState & ProcessingState::EverythingDone) && !strcmp(kextList[i].id, kextHSWFbId))
-        {
+        if (!(progressState & ProcessingState::EverythingDone) &&
+            memcmp(kextList[i].id, kextHSWFbId, strlen(kextHSWFbId)) == 0) {
             SYSLOG(kThisKextID, "found %s", kextList[i].id);
             // 10.10.1- do not require any internal display patches for 0x0a2e0008
             // but it's unclear that if 0x0a26000a and 0x0a2e000a require this
@@ -177,15 +176,15 @@ void LWEnabler::processKext(KernelPatcher& patcher, size_t index, mach_vm_addres
                         return;
                     }
                     SYSLOG(kThisKextID, "binary patches for internal display have been generated.");
-                    KextPatch azul_patch_info {
+                    KextPatch patchData {
                         { &kextList[i], find, repl, sizeof(find), 1 },
                         KernelVersion::Yosemite, KernelVersion::HighSierra
                     };
                     
-                    applyPatches(patcher, index, &azul_patch_info, 1);
+                    applyPatches(patcher, index, &patchData, 1);
                     SYSLOG(kThisKextID, "enable internal display after sleep for ig-platform-id: 0x%08x", gIgPlatformId);
                     progressState |= ProcessingState::EverythingDone;
-                    
+                    break;
                 } else {
                     SYSLOG(kThisKextID, "cannot find platform-id at %s", *kextHSWFb);
                     return;
@@ -195,25 +194,63 @@ void LWEnabler::processKext(KernelPatcher& patcher, size_t index, mach_vm_addres
                 return;
             }
         }
-        //
-        // Enable lid wake for Skylake (skl) platform
-        //
-        if (!(progressState & ProcessingState::EverythingDone) && !strcmp(kextList[i].id, kextSKLFbId))
-        {
-            SYSLOG(kThisKextID, "found %s", kextList[i].id);
-            // it must be 0x19260004 due to the previous isFixablePlatform checking
-            const uint8_t skl_find[] = { 0x0a, 0x0b, 0x03, 0x00, 0x00, 0x07, 0x06, 0x00, 0x03, 0x00, 0x00, 0x00, 0x04, 0x00, 0x00, 0x00 };
-            const uint8_t skl_repl[] = { 0x0f, 0x0b, 0x03, 0x00, 0x00, 0x07, 0x06, 0x00, 0x03, 0x00, 0x00, 0x00, 0x04, 0x00, 0x00, 0x00 };
-            KextPatch skl_patch_info {
-                { &kextList[i], skl_find, skl_repl, sizeof(skl_find), 1 },
-                KernelVersion::ElCapitan, KernelVersion::HighSierra
-            };
 
-            applyPatches(patcher, index, &skl_patch_info, 1);
-            SYSLOG(kThisKextID, "enable internal display after sleep for ig-platform-id: 0x%08x", gIgPlatformId);
-            
-            progressState |= ProcessingState::EverythingDone;
-            break;
+        // Enable lid wake for Skylake (skl) platform
+        if (!(progressState & ProcessingState::EverythingDone) &&
+            memcmp(kextList[i].id, kextSKLFbId, strlen(kextSKLFbId)) == 0) {
+            SYSLOG(kThisKextID, "found %s", kextList[i].id);
+            // it must be 0x19260004 due to the previous configPlatform's check
+            mach_vm_address_t address = patcher.solveSymbol(index,
+                                                            "__ZZN11BanksiaTcon10processCmdE22kFBControllerCommand_tPmmS1_S1_E14tconFeatureSet");
+            if (address) {
+                SYSLOG(kThisKextID, "obtained __ZZN11BanksiaTcon10processCmdE22kFBControllerCommand_tPmmS1_S1_E14tconFeatureSet");
+                patcher.clearError();
+                // Lookup the ig-platform-id specific framebuffer data
+                auto curOff = reinterpret_cast<uint8_t *>(address);
+                // The real patch place should be very close
+                // MaxSearchSize aka PAGE_SIZE is fairly enough
+                auto endOff = curOff + PAGE_SIZE;
+                // The framebuffer size
+                static constexpr size_t MaxReplSize {16};
+                // Search the specific ig-platform-id in the neighbourhood
+                while (curOff < endOff && memcmp(curOff, rIgPlatformId, sizeof(rIgPlatformId)))
+                    curOff++;
+                // verify search
+                if (curOff < endOff) {
+                    DBGLOG(kThisKextID, "found platform-id (0x%08x) at framebuffer info data segment", gIgPlatformId);
+                    // now let's generate the patch for it
+                    uint8_t find[MaxReplSize] {};
+                    uint8_t repl[MaxReplSize] {};
+                    lilu_os_memcpy(find, curOff+96, MaxReplSize);
+                    lilu_os_memcpy(repl, curOff+96, MaxReplSize);
+                    DBGLOG(kThisKextID, "%u, %u, %u, %u, %u", *find, *(find+1), *(find+2), *(find+3), *(find+4));
+                    // apply platform specific patch pattern
+                    // 0x19260004 uses 0x0f
+                    memset(repl+1, 0x0f, sizeof(uint8_t));
+                    if (memcmp(repl+1, find+1, sizeof(uint8_t)) == 0) {
+                        // already patch due to the kext has been
+                        // patched on the cache? we should stop here
+                        SYSLOG(kThisKextID, "already enabled internal display after sleep for ig-platform-id: 0x%08x", gIgPlatformId);
+                        return;
+                    }
+                    SYSLOG(kThisKextID, "binary patches for internal display have been generated.");
+                    KextPatch patchData {
+                        { &kextList[i], find, repl, sizeof(find), 1 },
+                        KernelVersion::ElCapitan, KernelVersion::HighSierra
+                    };
+                    
+                    applyPatches(patcher, index, &patchData, 1);
+                    SYSLOG(kThisKextID, "enable internal display after sleep for ig-platform-id: 0x%08x", gIgPlatformId);
+                    progressState |= ProcessingState::EverythingDone;
+                    break;
+                } else {
+                    SYSLOG(kThisKextID, "cannot find platform-id at %s", *kextHSWFb);
+                    return;
+                }
+            } else {
+                SYSLOG(kThisKextID, "cannot find __ZZN11BanksiaTcon10processCmdE22kFBControllerCommand_tPmmS1_S1_E14tconFeatureSet");
+                return;
+            }
         }
     }
     patcher.clearError();
@@ -229,7 +266,7 @@ void LWEnabler::applyPatches(KernelPatcher& patcher, size_t index, const KextPat
             if (patcher.compatibleKernel(patch.minKernel, patch.maxKernel))
             {
                 patcher.applyLookupPatch(&patch.patch);
-                SYSLOG(kThisKextID, "patch %s (%ld/%ld).", patch.patch.kext->id, p+1, patchNum);
+                DBGLOG(kThisKextID, "patch %s (%ld/%ld).", patch.patch.kext->id, p+1, patchNum);
                 patcher.clearError();
             }
         }